backend/requirements.txt
# Flask ve Extensions
Flask>=3.0.0
flask-cors>=4.0.0
flask-jwt-extended>=4.6.0
flask-migrate>=4.0.0
flask-restx>=1.3.0
flask-sqlalchemy>=3.1.0
Werkzeug>=3.0.0

# Database
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.0

# Redis
redis>=5.0.0

# Blockchain
web3>=6.0.0

# HTTP Client
requests>=2.31.0

# Environment
python-dotenv>=1.0.0

# Production Server
gunicorn>=21.0.0

# Development
pytest>=7.4.0
pytest-flask>=1.3.0
